Make parts of cli_wrapper test-only.
